# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "Contacter Server",
    "description": "API documentation"
  },
  "servers": [
    {
      "url": "https://localhost:3000/api"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contact/{method}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "method",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Auth_user": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contact": {
                    "type": "string",
                    "example": "jack@gmail.com"
                  },
                  "data": {
                    "type": "object",
                    "example": {
                      "subject": "Subject",
                      "html": "<h1>Hello Jack !</h1>"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/has_access/{method}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "method",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/methods": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "Auth_user": {
        "type": "apiKey",
        "description": "Admin or Voter auth token",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": {
    "Auth_user": []
  }
}